+++ python/eyes_robotframework/src/EyesLibrary/__init__.py
@@ -13,6 +13,7 @@ from applitools.common import BatchInfo
 from applitools.common.utils import argument_guard
 from applitools.selenium import ClassicRunner, Eyes, VisualGridRunner
 from applitools.selenium.eyes import EyesRunner
+from applitools.selenium.fluent import SeleniumCheckSettings
 
 from .__version__ import __version__
 from .config import RobotConfiguration
@@ -184,6 +185,7 @@ class EyesLibrary(DynamicCore):
         self._batch_registry = {}  # type: Dict[Text, BatchInfo]
         self._running_keyword = None
         self._configuration = None
+        self._current_check_settings = SeleniumCheckSettings()
 
         self._selected_runner = try_parse_runner(runner)
 
@@ -313,3 +315,10 @@ class EyesLibrary(DynamicCore):
     @property
     def configure(self):
         return self._configuration
+
+    @property
+    def current_check_settings(self):
+        return self._current_check_settings
+
+    def set_current_check_settings(self, check_settings):
+        self._current_check_settings = check_settings