+++ python/eyes_robotframework/src/EyesLibrary/keywords/target.py
@@ -39,7 +39,10 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=    | Target Window     |
         """
-        return collect_check_settings(Target.window(), *check_settings_keywords)
+        self.set_current_check_settings(Target.window())
+        return collect_check_settings(
+            self.current_check_settings, *check_settings_keywords
+        )
 
     @keyword(
         "Target Region By Element",
@@ -56,7 +59,10 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Region By Element  |  ${element}  |
         """
-        return collect_check_settings(Target.region(element), *check_settings_keywords)
+        self.set_current_check_settings(Target.region(element))
+        return collect_check_settings(
+            self.current_check_settings, *check_settings_keywords
+        )
 
     @keyword(
         "Target Region By Coordinates", types=(str,), tags=(CHECK_SETTINGS_SUPPORT,)
@@ -72,8 +78,9 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Region By Coordinates  |  [10 30 40 50]  |
         """
+        self.set_current_check_settings(Target.region(parse_region(region)))
         return collect_check_settings(
-            Target.region(parse_region(region)), *check_settings_keywords
+            self.current_check_settings, *check_settings_keywords
         )
 
     @keyword("Target Region By Selector", types=(str,), tags=(CHECK_SETTINGS_SUPPORT,))
@@ -88,9 +95,11 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Frame By Selector  |  css:#selector  |
         """
+        self.set_current_check_settings(
+            Target.region(self.from_locator_to_supported_form(selector))
+        )
         return collect_check_settings(
-            Target.region(self.from_locator_to_supported_form(selector)),
-            *check_settings_keywords
+            self.current_check_settings, *check_settings_keywords
         )
 
     @keyword(
@@ -111,9 +120,7 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Region By Target Path  | Shadow By Selector  | css:#has-shadow-root |  Region By Selector | //selector |
         """
-        return collect_check_settings_with_target_path(
-            Target.region, *check_settings_keywords
-        )
+        return collect_check_settings_with_target_path(self, *check_settings_keywords)
 
     @keyword(
         "Target Frame By Element",
@@ -131,7 +138,10 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Frame By Element  |  ${element}  |
         """
-        return collect_check_settings(Target.frame(element), *check_settings_keywords)
+        self.set_current_check_settings(Target.frame(element))
+        return collect_check_settings(
+            self.current_check_settings, *check_settings_keywords
+        )
 
     @keyword(
         "Target Frame By Selector",
@@ -149,9 +159,11 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Frame By Selector  |  css:#selector  |
         """
+        self.set_current_check_settings(
+            Target.frame(self.from_locator_to_supported_form(selector))
+        )
         return collect_check_settings(
-            Target.frame(self.from_locator_to_supported_form(selector)),
-            *check_settings_keywords
+            self.current_check_settings, *check_settings_keywords
         )
 
     @keyword(
@@ -170,8 +182,9 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             | ${target}=  |  Target Frame By Index  |  2  |
         """
+        self.set_current_check_settings(Target.frame(frame_index))
         return collect_check_settings(
-            Target.frame(frame_index), *check_settings_keywords
+            self.current_check_settings, *check_settings_keywords
         )
 
     @keyword(
@@ -190,6 +203,7 @@ class TargetKeywords(LibraryComponent):
         *Example:*
             |  ${target}=  |  Target Frame By Name  |  frameName  |
         """
+        self.set_current_check_settings(Target.frame(frame_name))
         return collect_check_settings(
-            Target.frame(frame_name), *check_settings_keywords
+            self.current_check_settings, *check_settings_keywords
         )